% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_term.R
\name{calc_month_term}
\alias{calc_month_term}
\title{calculate term by month}
\usage{
calc_month_term(min, max)
}
\arguments{
\item{min}{Date}

\item{max}{Date}
}
\value{
int how many terms
}
\description{
calculate how many terms by month.
}
\examples{
calc_quarter_term(as.Date("2001-01-01"),as.Date("2001-12-01"))
}
