% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_term.R
\name{calc_quarter_term}
\alias{calc_quarter_term}
\title{calculate term by quarter}
\usage{
calc_quarter_term(min, max)
}
\arguments{
\item{min}{Date}

\item{max}{Date}
}
\value{
int how many terms
}
\description{
calculate how many terms by quarter.
}
\examples{
calc_quarter_term(as.Date("2001-01-01"),as.Date("2001-12-01"))
}
