% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_month}
\alias{classify_month}
\title{classify by month}
\usage{
classify_month(ts)
}
\arguments{
\item{ts}{vector of Date}
}
\value{
vector of Date
}
\description{
make label for classify by month.
}
\examples{
classify_month(as.Date("2020-01-01"))
}
