% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_quarter}
\alias{classify_quarter}
\title{classify by quarter}
\usage{
classify_quarter(ts)
}
\arguments{
\item{ts}{vector of Date}
}
\value{
vector of Date
}
\description{
make label for classify by quarter.
}
\examples{
classify_quarter(as.Date("2020-01-01"))
}
