% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_year}
\alias{classify_year}
\title{classify by year}
\usage{
classify_year(ts)
}
\arguments{
\item{ts}{vector of Date}
}
\value{
vector of Date
}
\description{
make label for classify by year.
}
\examples{
classify_year(as.Date("2020-01-01"))
}
