% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{seq_date_by_month}
\alias{seq_date_by_month}
\title{make sequence date data by month}
\usage{
seq_date_by_month(min, max)
}
\arguments{
\item{min}{Date}

\item{max}{Date}
}
\value{
vector of Date
}
\description{
make sequence date data by month.
}
\examples{
seq_date_by_month(as.Date("2000-01-01"),as.Date("2010-01-01"))
}
