% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{seq_date_by_quarter}
\alias{seq_date_by_quarter}
\title{make sequence date data by quarter}
\usage{
seq_date_by_quarter(min, max)
}
\arguments{
\item{min}{Date}

\item{max}{Date}
}
\value{
vector of Date
}
\description{
make sequence date data by quarter.
}
\examples{
seq_date_by_quarter(as.Date("2000-01-01"),as.Date("2010-01-01"))
}
